Move types and mixin to sensor
